requirements.txt
pyttsx3
gtts
pygame
openai
SpeechRecognition
pyaudio
sounddevice
pillow
tk